+++ client/src/pages/QuizOverviewPage.tsx
@@ -1,4 +1,4 @@
-import QuizOverview from "../components/QuizOverview";
+import QuizOverview from "../components/quiz/QuizOverview";
 import { useRequireAuth } from "../hooks/useRequireAuth";
 
 export default function QuizOverviewPage() {
